extracted classes, added icon tooltips
