+++ Rubberduck.Core/UI/FindSymbol/FindSymbolControl.xaml
@@ -10,9 +10,17 @@
              d:DataContext="{d:DesignInstance {x:Type local:FindSymbolViewModel}, IsDesignTimeCreatable=False}">
 
     <UserControl.Resources>
-        <BitmapImage x:Key="ArrowImage" UriSource="pack://application:,,,/Rubberduck.Resources;component/Icons/Fugue/arrow.png" />
-        <local:SearchBoxMultiBindingConverter x:Key="SearchBoxTextConverter" />
-        <converters:DeclarationToIconConverter x:Key="IconConverter" />
+        <ResourceDictionary>
+            <ResourceDictionary.MergedDictionaries>
+                <ResourceDictionary Source="../Controls/ToolBar.xaml"/>
+                <ResourceDictionary Source="../Styles/DefaultStyle.xaml"/>
+            </ResourceDictionary.MergedDictionaries>
+            <BitmapImage x:Key="ArrowImage" UriSource="pack://application:,,,/Rubberduck.Resources;component/Icons/Fugue/arrow.png" />
+            <local:SearchBoxMultiBindingConverter x:Key="SearchBoxTextConverter" />
+            <converters:DeclarationToIconConverter x:Key="IconConverter" />
+            <converters:AccessibilityToIconConverter x:Key="AccessibilityToIcon" />
+            <converters:DeclarationToDeclarationTypeStringConverter x:Key="DeclarationTypeConverter" />
+        </ResourceDictionary>
     </UserControl.Resources>
     
     <Grid>
@@ -46,10 +54,13 @@
                             <ColumnDefinition MinWidth="192" />
                             <ColumnDefinition Width="*" />
                         </Grid.ColumnDefinitions>
-                        <Image Grid.Column="0" Height="16" Width="16" Margin="2" HorizontalAlignment="Center"
+                        <Image Grid.Column="0" Style="{StaticResource ToolbarIconStyle}" HorizontalAlignment="Center"
                                Source="{Binding Declaration, Converter={StaticResource IconConverter}, Mode=OneTime}" 
-                               ToolTip="{Binding Declaration.DeclarationType}" />
-                        <TextBlock Grid.Column="1" Text="{Binding IdentifierName}" FontWeight="Bold" MinWidth="160" VerticalAlignment="Center" />
+                               ToolTip="{Binding Declaration, Converter={StaticResource DeclarationTypeConverter}, Mode=OneTime}" />
+                        <Image Grid.Column="0" Style="{StaticResource ToolbarIconStyle}" HorizontalAlignment="Center"
+                               Source="{Binding Declaration, Converter={StaticResource AccessibilityToIcon}}"
+                               ToolTip="{Binding Declaration, Converter={StaticResource DeclarationTypeConverter}, Mode=OneTime}" />
+                        <TextBlock Grid.Column="1" Text="{Binding IdentifierName}" FontWeight="Bold" FontFamily="Consolas" MinWidth="160" VerticalAlignment="Center" />
                         <TextBlock Grid.Column="2" Text="{Binding Location}" Margin="20,0" Foreground="DimGray" FontStyle="Italic" VerticalAlignment="Center" HorizontalAlignment="Right" />
                     </Grid>
                 </DataTemplate>
